added base class for jobstatuscache
